Fix rule arg type inferencing
The type checker was not inferring the types of function args correctly
in all cases. For example:

f(x) { [y] = x } # ok
f(x) { x = [y] } # error

In this case, it would infer x/array<any> and y/any in the first case
but the second would error. This was because the unify check was broken.

Fixes #542