Fix rule graph construction for rule's else chain
The rule graph needs to have edges for all rules in the chain (not just
the first most one.) This was resulting in false-postiive errors in the
type checker when rule arg types should have unioned, e.g.:

f(x) {
    x = 1
    # some other expr
} else {
    x = []
    # some other expr
} else {
    x = "foo"
}

p { f("foo") }

In cases like this, p may have been type checked before the entire f
chain. As a result, it could fail if the last rule in the chain had not
been checked.