Update QueryCompiler to report rewritten vars
Since assigned vars are rewritten, we need to allow callers to lookup
whether a generated var corresponds to a var in the parsed query.
Previously, no generated vars needed to be reported to the callers of
the rego package (and so they are filtered out from the ResultSet). Now,
that generated vars correspond to vars the caller may care about, the
rego package has to map generated vars back to ones the user knows
about.