Fix partial evaluation of refs on unknown terms
Previously, if partial evaluation encountered a ref against an unknown
term, it would attempt to evaluate the ref against the binding list and
then fail (because there would not be a binding for the unknown.)

With thes changes partial evaluation deals with unknown terms twofold:

1) When the term being dereferenced is unknown, eval saves the remainder
of the expression and continues.

2) When the ref operand is unknown, eval uses the value obtained by
enumerating the term to continue.