Plug comprehensions in partial eval results
During partial evaluation, comprehensions were not being plugged before
being returned to callers. If the comprehension closed over a variable
that was not included in the result, the comprehension would become
unsafe.

With this change, comprehensions are plugged before being returned to
callers. This will prevent these variables from becoming unsafe.

In the future, when comprehensions are evaluated during partial
evaluation, this may be unnecessary.

Fixes #656

Signed-off-by: Torin Sandall <torinsandall@gmail.com>