Improve partial evaluation testing
These changes update the topdown test suite to run tests using normal
evaluation and then again against the result of partial evaluation. In
the second pass, input is treated as unknown.

This additional coverage uncovered an issue in the namespacing applied
before partial evaluation results are returned to callers. Specifically,
reference heads and with values were not being namespaced.

The http.send POST test was also updated to avoid keeping track of
request bodies as this broke when the case was executed twice.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>