Set bundle status in case of HTTP 304
Previously, the bundle plugin was only updating the status struct if:

* a download or activation error occurred.
* a new bundle was successfully activated.

As a result, if a transient bundle download/activation error occcured,
OPA would never report that the error condition had been cleared in the
event that the service began serving HTTP 304s.

These changes update the bundle plugin to always set the error status
regardles of whether an error occurred or a new bundle was activated.

Also, refactor the bundle plugin to set all of the activation state
(timestamp, revision, etag) in one place.

Fixes #794

Signed-off-by: Torin Sandall <torinsandall@gmail.com>