Add Partial function to Rego objects
Partial allows callers to partially evaluate a query and obtain the
resulting partially evaluated queries and support modules from the
evaluation engine.

This is a lower level interface compared to the other API (PartialEval)
that returns a result which can be evaluated normally (which takes
advantage of partial evaluation for optimization purposes.) This new
interface is targetted at use cases where callers are not interested in
a binary policy decision but instead need the conditions that should be
evaluated at a later time.

Deprecate the old PartialEval function and rename it to PartialResult to
avoid some (but certainly not all) confusion.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>