Filled out the casting section of builtin functions. I think the to_set and to_array functions require no justification.
For the others: I realized this would be useful to have when I was
doing http calls that returned a types.A and thus when I parsed my code
on occasion it only threw type errors at runtime when I attempted to use
the result in another builtin. It would be great to have some way to
throw these type errors at compile time as well if the
user likes, hence these functions.

Changed names to cast_, added test, added copyright banner, added docs.

Signed-off-by: Varun Mathur <varun.mathur@live.com>
Signed-off-by: Varun Mathur <vmathur@cloudsimple.com>
Signed-off-by: Varun Mathur <varun.mathur@live.com>