Refactor save set updating to use ast.Walk
In 5b5ebf7 we fixed an issue in partial eval that was causing vars
embedded inside composites to not be added to the save set when they
should have been.

This change simply refactors that fix to use ast.Walk and rely on the
same logic in both saveUnify and saveCall.

Note, the plug call on the var terms has been removed as it's not needed
since we are already plugging terms (or applying) when checking if terms
are contained in the save set.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>