Remove support rule if default value is not needed
Previously, partial eval would always generate a support rule when it
evaluated a rule with a default value. These changes tweak this
behaviour so that support rules are only generated if the default value
is required in the partial eval output. It's better to avoid generating
support rules as it simplifies the assumptions that the consumer of
partial eval can make.

Fixes #820

Signed-off-by: Torin Sandall <torinsandall@gmail.com>