Improve parser error readability
PreviousLy if a match failure occurred, the error would be returned to
the caller as is. Because of how the grammar is specified this often
resulted in noisy error messages that list all of the low level symbols
that the parser _tried_ to match (e.g., #, %, /, >, etc.) The expected
values are basically useless for users.

With these changes the expected values are removed from the parser
errors and the parser errors now include the text from the line in
question and a carrot identifying where the failure occurred.

These changes also improve the error hanlding for non-terminated strings
and other well-known classes of parser errors. In the future we can
improve the parser to return more specific errors that are more useful
than the "no match found" default.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>