Improve Docker tutorial example
Previously the example policy at the end of the tutorial would allow
operations if the Authz-User was contained in the input headers but not
in the user mapping. This change makes sure that the user is defined in
the mapping.

Also, update the example to use := and == instead of = (which is
functionaly the same but better practice because they avoid capturing
globals.)

Fixes #880

Signed-off-by: Torin Sandall <torinsandall@gmail.com>