Removing FOSSA badge from README
The FOSSA license scan recently began failing for some (unknown) reason.
Since the vendoring has not been updated lately, it seems like this is
not a valid warning. FOSSA has not responded to support requests for
several days now, so I removing the badge until we can get the issue
reoslved.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>