Fix copy propagator type assertion panic
The copy propagator was not taking the union-find subsitutions into
account when building the headvar set. As a result, bindings were being
added for call expressions when the var would later be substituted for a
ref head (which ultimately would cause a panic when the ref head was
substituted for the call).

These changes update the copy propagator to add the union-find root to
the headvar set instead of the headvar. This way the headvar
construction is consistent with the binding plug and update below.

Fixes #912

Signed-off-by: Torin Sandall <torinsandall@gmail.com>