Apply copy propagation to support rules
Previously copy propagation was only being applied to queries returned
by partial eval. The assumption was that clients would typically be
implemented to deal with queries and that support rules would be
unhandled. In some cases though, clients may be capable of handling
support rules. For them, it's better if we apply copy propagation to
both queries and rule bodies.

Note, we ensure that rule bodies are non-empty so that comparisons to
parsed rules that are equivalent (but lacking a body) succeed.

Fixes #911

Signed-off-by: Torin Sandall <torinsandall@gmail.com>