Update copy propagation to kill single term exprs
These changes update copy propagation to kill single term expressions
defined by vars that are in the binding list. Previously the term would
be substituted with the binding value which results in unnecessary
expressions being returned.

Note without a type environment we cannot remove all single term
expressions as copy propagation does not know if refs refer to sets or
arrays/objects. Single term expressions that refer to sets could be
killed if the type information was available.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>