Fix cross-compile issue in go generate step
The WASM changes broke cross-compiling because go generate executes go
run ... which inherits the GOOS and GOARCH environment variables. To
resolve this we just wrap the go rnu ... command in a script that
overwrites the environment variables so that the host GOOS and GOARCH
are used. For some reason go generate does not allow you to specify
environment variables inline, e.g., FOO=BAR go run ...

Signed-off-by: Torin Sandall <torinsandall@gmail.com>