Fix bug in test runner rule name dedup
The test runner was using the rule name as the cache key for
deduplication purposes. If test rules in two different packages had the
same name, the second one would be rewritten which could cause
unexpected errors.

These changes simply update the runner to use the rule path (which is
the package path + rule name) because that ought to be unique.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>