Fix partial eval to apply saved terms
Previously we were not applying the bindings to vars before adding them
to the save set. This meant the biunifier and the save set were out of
sync, e.g., a the biunifier would have (w/(x,1), 2) while the save set
had (w, 2). Therefore, when an expression like w = y was applied, the
result would be x = y which would not be saved. These changes modify
evaluation to apply vars before adding them to the save set. This way
the save set would correctly have (x, 1) in the case above.

Also, add String() function to save set for future debugging.

Fixes #1074

Signed-off-by: Torin Sandall <torinsandall@gmail.com>